refactor: extract re-usable BrandedNavBar components
